Move ManageSubscriptionsCallback to the interfaces package
Every other callback handed to a Purchases method lives in
com.revenuecat.purchases.interfaces (17 of the 19 public callback and
listener interfaces in this module). The exceptions are
CustomerCenterListener, which sits with its own feature package, and
LogHandler. UpdatedCustomerInfoListener shows that even settable
listeners belong in interfaces, so there is no carve out for this one.

Doing it now because it is free. Once shipped in the root package, the
move would be a source and binary break (the JVM encodes the package in
the class name, so already compiled apps would hit NoSuchMethodError on
the showManageSubscriptions call site) and a typealias would not help
Java consumers.

Signature files regenerated with scripts/api-dump.sh.

Reported by Cursor Bugbot.
