+++ purchases/api-defauts.txt
@@ -213,11 +213,6 @@ package com.revenuecat.purchases {
   public static final class LogLevel.Companion {
   }
 
-  public interface ManageSubscriptionsCallback {
-    method public void onError(com.revenuecat.purchases.PurchasesError error);
-    method public void onSuccess();
-  }
-
   public final class NoCoreLibraryDesugaringException extends java.lang.RuntimeException {
     ctor public NoCoreLibraryDesugaringException(Throwable cause);
   }
@@ -474,7 +469,7 @@ package com.revenuecat.purchases {
     method public void showInAppMessagesIfNeeded(android.app.Activity activity);
     method public void showInAppMessagesIfNeeded(android.app.Activity activity, optional java.util.List<? extends com.revenuecat.purchases.models.InAppMessageType> inAppMessageTypes);
     method public void showManageSubscriptions(android.content.Context context);
-    method public void showManageSubscriptions(android.content.Context context, optional com.revenuecat.purchases.ManageSubscriptionsCallback? callback);
+    method public void showManageSubscriptions(android.content.Context context, optional com.revenuecat.purchases.interfaces.ManageSubscriptionsCallback? callback);
     method @Deprecated public void syncAmazonPurchase(String productID, String receiptID, String amazonUserID, String? isoCurrencyCode, Double? price);
     method public void syncAmazonPurchase(String productID, String receiptID, String amazonUserID, String? isoCurrencyCode, Double? price, long purchaseTime);
     method public void syncAttributesAndOfferingsIfNeeded(com.revenuecat.purchases.interfaces.SyncAttributesAndOfferingsCallback callback);
@@ -1131,6 +1126,11 @@ package com.revenuecat.purchases.interfaces {
   public interface LogInCallback {
     method public void onError(com.revenuecat.purchases.PurchasesError error);
     method public void onReceived(com.revenuecat.purchases.CustomerInfo customerInfo, boolean created);
+  }
+
+  public interface ManageSubscriptionsCallback {
+    method public void onError(com.revenuecat.purchases.PurchasesError error);
+    method public void onSuccess();
   }
 
   @com.revenuecat.purchases.ExperimentalPreviewRevenueCatPurchasesAPI public fun interface PollRewardVerificationCallback {
